Add tests

+++ src/Shell/PhpstormShell.php
@@ -13,8 +13,6 @@ use IdeHelper\Generator\TaskCollection;
  */
 class PhpstormShell extends Shell {
 
-	const CODE_CHANGES = 2;
-
 	/**
 	 * @return void
 	 */
@@ -28,24 +26,27 @@ class PhpstormShell extends Shell {
 	 * @return int
 	 */
 	public function generate() {
-		$taskCollection = new TaskCollection();
-		$phpstormGenerator = new PhpstormGenerator($taskCollection);
-
+		$phpstormGenerator = $this->getGenerator();
 		$content = $phpstormGenerator->generate();
 
-		$file = ROOT . DS . '.phpstorm.meta.php';
+		$file = $this->getMetaFilePath();
 
-		$currentContent = file_get_contents($file);
+		$currentContent = file_exists($file) ? file_get_contents($file) : null;
 		if ($content === $currentContent) {
 			$this->out('Meta file `.phpstorm.meta.php` still up to date.');
 			return parent::CODE_SUCCESS;
+		}
+
+		if ($this->param('dry-run')) {
+			$this->out('Meta file `.phpstorm.meta.php` needs updating.');
+			return static::CODE_ERROR;
 		}
 
 		file_put_contents($file, $content);
 
 		$this->out('Meta file `.phpstorm.meta.php` generated.');
 
-		return static::CODE_CHANGES;
+		return static::CODE_SUCCESS;
 	}
 
 	/**
@@ -53,6 +54,13 @@ class PhpstormShell extends Shell {
 	 */
 	public function getOptionParser() {
 		$subcommandParser = [
+			'options' => [
+				'dry-run' => [
+					'short' => 'd',
+					'help' => 'Dry run the task. This will output an error code 1 if file needs changing. Can be used for CI checking.',
+					'boolean' => true,
+				],
+			]
 		];
 
 		return parent::getOptionParser()
@@ -61,6 +69,22 @@ class PhpstormShell extends Shell {
 				'help' => 'Generate `.phpstorm.meta.php` meta file.',
 				'parser' => $subcommandParser
 			]);
+	}
+
+	/**
+	 * @return \IdeHelper\Generator\PhpstormGenerator
+	 */
+	protected function getGenerator() {
+		$taskCollection = new TaskCollection();
+
+		return new PhpstormGenerator($taskCollection);
+	}
+
+	/**
+	 * @return string
+	 */
+	protected function getMetaFilePath() {
+		return ROOT . DS . '.phpstorm.meta.php';
 	}
 
 }
